Fix issue #5
Position of x ticks is set relatively to the left edge of the bar
groups.  The tick was always being placed just 1 bar width to the right,
resulting in asymetrical placement of the tick in case when bar groups
were wider than 2*width.  Now the shift equals group_width/2 that makes
bar groups centered around x tick for any number of datasets.

The space on the right of the last bar was slightly wider than it was
supposed to be (because it was calculated incorrectly).  This was also
fixed in this commit.